# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: taboo · 禁制 — 密钥管理平台 API
  version: 0.7.0
  description: 开源自部署密钥管理平台。高价值数据 + 严格访问控制 + 全程留痕。M5：Secret Sync / Webhooks / OIDC SSO / 审计导出。
servers:
  - url: http://localhost:7100
    description: 本地开发
...
paths:
...
  /api/v1/orgs/{slug}/members:
    get:
      tags: [members]
      summary: 成员列表（owner）
      operationId: listMembers
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgSlug' }]
      responses:
        '200':
          description: 成员
          content:
            application/json:
              schema:
                type: object
                properties:
                  members: { type: array, items: { $ref: '#/components/schemas/OrgMember' } }
    post:
      tags: [members]
      summary: 添加已注册用户为成员
      operationId: addMember
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgSlug' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
                role: { type: string, enum: [owner, admin, developer, viewer] }
                restricted: { type: boolean }
      responses:
        '201': { description: 已添加 }
  /api/v1/orgs/{slug}/members/{userId}:
    patch:
      tags: [members]
      summary: 改角色 / restricted
      operationId: patchMember
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [owner, admin, developer, viewer] }
                restricted: { type: boolean }
      responses:
        '200': { description: 已更新 }
    delete:
      tags: [members]
      summary: 移除成员
      operationId: removeMember
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 已移除 }
  /api/v1/orgs/{slug}/members/{userId}/grants:
    put:
      tags: [members]
      summary: 替换项目授权矩阵
      operationId: putMemberGrants
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                grants:
                  type: array
                  items:
                    type: object
                    properties:
                      project_id: { type: string }
                      role: { type: string }
      responses:
        '200': { description: 已更新 }
  /api/v1/orgs/{slug}/invites:
    post:
      tags: [members]
      summary: 创建邀请（自托管返回一次性链接）
      operationId: createInvite
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgSlug' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
                role: { type: string, enum: [owner, admin, developer, viewer] }
      responses:
        '201':
          description: 邀请链接
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  url: { type: string }
                  token: { type: string }
                  expires_at: { type: integer }
  /api/v1/invites/{token}/accept:
    post:
      tags: [members]
      summary: 接受邀请（登录用户邮箱须匹配）
      operationId: acceptInvite
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 已加入 }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    OrgSlug: { name: slug, in: path, required: true, schema: { type: string } }
...
  schemas:
...
    OrgMember:
      type: object
      properties:
        user_id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [owner, admin, developer, viewer] }
        restricted: { type: boolean }
        joined_at: { type: string }
        grants:
          type: array
          items:
            type: object
            properties:
              project_id: { type: string }
              project_slug: { type: string }
              project_name: { type: string }
              role: { type: string }
...